+++ src/base/Global.h
@@ -55,8 +55,8 @@ void SetTraceFile(HANDLE TraceFile);
 void CleanupTracing();
 #define TRACEENV "WINSCPTRACE"
 extern BOOL IsTracing;
-const unsigned int CallstackTlsOff = (unsigned int)-1;
-extern unsigned int CallstackTls;
+const uintptr_t CallstackTlsOff = (uintptr_t)-1;
+extern uintptr_t CallstackTls;
 extern "C" void DoTrace(const wchar_t * SourceFile, const wchar_t * Func,
   intptr_t Line, const wchar_t * Message);
 void DoTraceFmt(const wchar_t * SourceFile, const wchar_t * Func,
